apispec Update monitoring tests fixtures.
